Fix problem with boost += assignment (#1)
Since boost 1.69.00, the expression
```
a += b;
```
where a and b are std::vector<std::string> does no longer
compile.

The documentation of operator+= in boost assign says that
the right hand expression must be convertable to value_type of the
underlying container. Not sure why this compiled previously with
b being a std::vector.

The patch brought forward here is to replace all
occurences of relevant += by a function call to a `vectorAppend`
function. This should work regardless of the boost version.